+++ backend/src/main/java/team/teamby/teambyteam/feed/presentation/FeedThreadController.java
@@ -12,6 +12,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import team.teamby.teambyteam.feed.application.FeedThreadService;
+import team.teamby.teambyteam.feed.application.FeedWriteService;
 import team.teamby.teambyteam.feed.application.dto.FeedThreadWritingRequest;
 import team.teamby.teambyteam.feed.application.dto.FeedsResponse;
 import team.teamby.teambyteam.member.configuration.AuthPrincipal;
@@ -23,6 +24,7 @@ import team.teamby.teambyteam.member.configuration.dto.MemberEmailDto;
 public class FeedThreadController {
 
     private final FeedThreadService feedThreadService;
+    private final FeedWriteService feedWriteService;
 
     @PostMapping("/{teamPlaceId}/feed/threads")
     public ResponseEntity<Void> write(
@@ -30,7 +32,7 @@ public class FeedThreadController {
             @AuthPrincipal final MemberEmailDto memberEmailDto,
             @PathVariable final Long teamPlaceId
     ) {
-        final Long threadId = feedThreadService.write(request, memberEmailDto, teamPlaceId);
+        final Long threadId = feedWriteService.write(request, memberEmailDto, teamPlaceId);
         final URI location = URI.create("/api/team-place/" + teamPlaceId + "/feed/threads/" + threadId);
 
         return ResponseEntity.created(location).build();
